2021_aircleaner/requirements.txt
pybluez
pygatt
tk
influxdb